+++ src/Soluble/Japha/Util/Exception/InvalidArgumentException.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * Soluble Japha
  *
@@ -10,9 +12,8 @@
 
 namespace Soluble\Japha\Util\Exception;
 
-use Soluble\Japha\Bridge\Exception\ExceptionInterface;
 use Soluble\Japha\Bridge\Exception\InvalidArgumentException as IAE;
 
-class InvalidArgumentException extends IAE implements ExceptionInterface
+class InvalidArgumentException extends IAE
 {
 }