+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 2.0.0 (to be release)
+
+## Changed
+
+- Support for PHP7.1 only `declare(strict_types=1)`
+
+## Fixed
+
+- Minor bugfix with integer cookie encoding due to strict_types
+
 ## 1.4.5 (2017-05-29)
 
 ## Fixed